Removed `guard x >= 0` from `pow(x, y)` implementations
It is perfectly reasonable to call `pow` with a negative base. If the exponent is an integer the calculation will succeed, and if not then the result of the `libm_pow` call will correctly be `.nan`. We don’t need to prohibit the valid case when we can simply call the library function unconditionally.