Remove anything that depends on differentiable programming
This turns out to not be fully upstreamed such that it doesn't work correctly in 5.2.x on Linux. We'll keep it around on the swift-5.3 branch (because the Swift toolchain there has support).

Looking ahead, I'd like to do any feature work that depends on unreleased Swift toolchain support on a "future" branch.